requirements.txt
flask
python-dotenv
openai
reportlab
flask-cors
